Split classes of txpool into different files
