+++ src/bigbang/entry.cpp
@@ -21,7 +21,8 @@
 #include "rpcclient.h"
 #include "rpcmod.h"
 #include "service.h"
-#include "txpool.h"
+#include "txpool/txpool.h"
+#include "txpool/txpoolcontroller.h"
 #include "version.h"
 #include "wallet.h"
 #include "worldline.h"
