+++ src/bigbang/txpool/txpoolview.h
@@ -1,0 +1,161 @@
+// Copyright (c) 2019 The Bigbang developers
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BIGBANG_TXPOOL_TXPOOLVIEW_H
+#define BIGBANG_TXPOOL_TXPOOLVIEW_H
+
+#include <map>
+
+#include "transaction.h"
+#include "txpool/txset.h"
+
+namespace bigbang
+{
+
+class CTxPoolView
+{
+public:
+    class CSpent : public CTxOut
+    {
+    public:
+        CSpent()
+          : txidNextTx(uint64(0)) {}
+        CSpent(const CTxOut& output)
+          : CTxOut(output), txidNextTx(uint64(0)) {}
+        CSpent(const uint256& txidNextTxIn)
+          : txidNextTx(txidNextTxIn) {}
+        void SetSpent(const uint256& txidNextTxIn)
+        {
+            *this = CSpent(txidNextTxIn);
+        }
+        void SetUnspent(const CTxOut& output)
+        {
+            *this = CSpent(output);
+        }
+        bool IsSpent() const
+        {
+            return (txidNextTx != 0);
+        };
+
+    public:
+        uint256 txidNextTx;
+    };
+
+public:
+    std::size_t Count() const
+    {
+        return setTxLinkIndex.size();
+    }
+    bool Exists(const uint256& txid) const
+    {
+        return setTxLinkIndex.count(txid) != 0;
+    }
+    CPooledTx* Get(uint256 txid)
+    {
+        CPooledTxLinkSetByTxHash& idxTx = setTxLinkIndex.get<0>();
+        CPooledTxLinkSetByTxHash::iterator mi = idxTx.find(txid);
+        if (mi == idxTx.end())
+        {
+            return nullptr;
+        }
+        return (*mi).ptx;
+    }
+    bool IsSpent(const CTxOutPoint& out) const
+    {
+        std::map<CTxOutPoint, CSpent>::const_iterator it = mapSpent.find(out);
+        if (it != mapSpent.end())
+        {
+            return (*it).second.IsSpent();
+        }
+        return false;
+    }
+    bool GetUnspent(const CTxOutPoint& out, CTxOut& unspent) const
+    {
+        std::map<CTxOutPoint, CSpent>::const_iterator it = mapSpent.find(out);
+        if (it != mapSpent.end() && !(*it).second.IsSpent())
+        {
+            unspent = static_cast<CTxOut>((*it).second);
+            return (!unspent.IsNull());
+        }
+        return false;
+    }
+    bool GetSpent(const CTxOutPoint& out, uint256& txidNextTxRet) const
+    {
+        std::map<CTxOutPoint, CSpent>::const_iterator it = mapSpent.find(out);
+        if (it != mapSpent.end())
+        {
+            txidNextTxRet = (*it).second.txidNextTx;
+            return (*it).second.IsSpent();
+        }
+        return false;
+    }
+    void SetUnspent(const CTxOutPoint& out)
+    {
+        CPooledTx* pTx = Get(out.hash);
+        if (pTx != nullptr)
+        {
+            mapSpent[out].SetUnspent(pTx->GetOutput(out.n));
+        }
+        else
+        {
+            mapSpent.erase(out);
+        }
+    }
+    void SetSpent(const CTxOutPoint& out, const uint256& txidNextTxIn)
+    {
+        mapSpent[out].SetSpent(txidNextTxIn);
+    }
+    void AddNew(const uint256& txid, CPooledTx& tx)
+    {
+        CPooledTxLinkSetByTxHash& idxTx = setTxLinkIndex.get<0>();
+        CPooledTxLinkSetByTxHash::iterator mi = idxTx.find(txid);
+        if (mi != idxTx.end())
+        {
+            setTxLinkIndex.erase(txid);
+        }
+        setTxLinkIndex.insert(CPooledTxLink(&tx));
+        for (std::size_t i = 0; i < tx.vInput.size(); i++)
+        {
+            mapSpent[tx.vInput[i].prevout].SetSpent(txid);
+        }
+        CTxOut output;
+        output = tx.GetOutput(0);
+        if (!output.IsNull())
+        {
+            mapSpent[CTxOutPoint(txid, 0)].SetUnspent(output);
+        }
+        output = tx.GetOutput(1);
+        if (!output.IsNull())
+        {
+            mapSpent[CTxOutPoint(txid, 1)].SetUnspent(output);
+        }
+    }
+    void Remove(const uint256& txid)
+    {
+        CPooledTx* pTx = Get(txid);
+        if (pTx != nullptr)
+        {
+            for (std::size_t i = 0; i < pTx->vInput.size(); i++)
+            {
+                SetUnspent(pTx->vInput[i].prevout);
+            }
+            setTxLinkIndex.erase(txid);
+        }
+    }
+    void Clear()
+    {
+        setTxLinkIndex.clear();
+        mapSpent.clear();
+    }
+    void InvalidateSpent(const CTxOutPoint& out, std::vector<uint256>& vInvolvedTx);
+    void ArrangeBlockTx(std::vector<CTransaction>& vtx, int64& nTotalTxFee, int64 nBlockTime, std::size_t nMaxSize) const;
+
+public:
+    CPooledTxLinkSet setTxLinkIndex;
+    std::map<CTxOutPoint, CSpent> mapSpent;
+};
+
+} // namespace bigbang
+
+#endif //BIGBANG_TXPOOL_TXPOOLVIEW_H
