+++ src/bigbang/txpool/txset.h
@@ -1,0 +1,75 @@
+// Copyright (c) 2019 The Bigbang developers
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BIGBANG_TXPOOL_TXSET_H
+#define BIGBANG_TXPOOL_TXSET_H
+
+#include <boost/multi_index/member.hpp>
+#include <boost/multi_index/ordered_index.hpp>
+#include <boost/multi_index_container.hpp>
+
+#include "transaction.h"
+
+namespace bigbang
+{
+
+class CPooledTx : public CAssembledTx
+{
+public:
+    std::size_t nSequenceNumber;
+    std::size_t nSerializeSize;
+
+public:
+    CPooledTx()
+    {
+        SetNull();
+    }
+    CPooledTx(const CAssembledTx& tx, std::size_t nSequenceNumberIn)
+      : CAssembledTx(tx), nSequenceNumber(nSequenceNumberIn)
+    {
+        nSerializeSize = xengine::GetSerializeSize(static_cast<const CTransaction&>(tx));
+    }
+    CPooledTx(const CTransaction& tx, int nBlockHeightIn, std::size_t nSequenceNumberIn, const CDestination& destInIn = CDestination(), int64 nValueInIn = 0)
+      : CAssembledTx(tx, nBlockHeightIn, destInIn, nValueInIn), nSequenceNumber(nSequenceNumberIn)
+    {
+        nSerializeSize = xengine::GetSerializeSize(tx);
+    }
+    void SetNull() override
+    {
+        CAssembledTx::SetNull();
+        nSequenceNumber = 0;
+        nSerializeSize = 0;
+    }
+};
+
+class CPooledTxLink
+{
+public:
+    CPooledTxLink()
+      : nSequenceNumber(0), ptx(nullptr) {}
+    CPooledTxLink(CPooledTx* ptxin)
+      : ptx(ptxin)
+    {
+        hashTX = ptx->GetHash();
+        nSequenceNumber = ptx->nSequenceNumber;
+    }
+
+public:
+    uint256 hashTX;
+    std::size_t nSequenceNumber;
+    CPooledTx* ptx;
+};
+
+typedef boost::multi_index_container<
+    CPooledTxLink,
+    boost::multi_index::indexed_by<
+        boost::multi_index::ordered_unique<boost::multi_index::member<CPooledTxLink, uint256, &CPooledTxLink::hashTX>>,
+        boost::multi_index::ordered_non_unique<boost::multi_index::member<CPooledTxLink, std::size_t, &CPooledTxLink::nSequenceNumber>>>>
+    CPooledTxLinkSet;
+typedef CPooledTxLinkSet::nth_index<0>::type CPooledTxLinkSetByTxHash;
+typedef CPooledTxLinkSet::nth_index<1>::type CPooledTxLinkSetBySequenceNumber;
+
+} // namespace bigbang
+
+#endif //BIGBANG_TXPOOL_TXSET_H
